requirements.txt
# Core dependencies
chromadb==1.0.7
sentence-transformers==4.1.0
groq==0.25.0
pdfplumber==0.11.6
python-dotenv==1.1.0

# Testing dependencies
pytest==8.3.5
hypothesis==6.131.0
